Start multi page app
