+++ .idea/sans_ml_streamlit.iml
@@ -5,4 +5,7 @@
     <orderEntry type="jdk" jdkName="Python 3.10 (deeplearning)" jdkType="Python SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
   </component>
+  <component name="PyDocumentationSettings">
+    <option name="renderExternalDocumentation" value="true" />
+  </component>
 </module>
